+++ Kiosk/Auction Listings/ListingsViewController.swift
@@ -120,6 +120,7 @@ class ListingsViewController: UIViewController {
 
         if let internalNav:FulfillmentNavigationController = containerController.internalNavigationController() {
             let registerVC = RegisterViewController.instantiateFromStoryboard()
+            registerVC.createNewUser = true
             internalNav.viewControllers = [registerVC]
         }
 
@@ -127,6 +128,10 @@ class ListingsViewController: UIViewController {
             containerController.viewDidAppearAnimation(containerController.allowAnimations)
         }
 
+    }
+    
+    @IBAction func longPressForAdmin(sender: AnyObject) {
+        self.performSegue(.ShowAdminOptions)
     }
     
     override func viewWillAppear(animated: Bool) {