+++ Kiosk/Bid Fulfillment/RegisterViewController.swift
@@ -11,9 +11,15 @@ class RegisterViewController: UIViewController {
 
     @IBOutlet var flowView: RegisterFlowView!
     @IBOutlet var bidDetailsPreviewView: BidDetailsPreviewView!
+    @IBOutlet var confirmButton: UIButton!
 
     let coordinator = RegistrationCoordinator()
+    let registrationNetworkModel = RegistrationNetworkModel()
 
+    dynamic var placingBid = false
+    dynamic var createNewUser = false
+    dynamic var details:BidDetails!
+    
     class func instantiateFromStoryboard() -> RegisterViewController {
         return UIStoryboard.fulfillment().viewControllerWithID(.RegisterAnAccount) as RegisterViewController
     }
@@ -24,8 +30,17 @@ class RegisterViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        
+        let registerIndexSignal = RACObserve(coordinator, "currentIndex")
+        let indexIsConfirmSignal = registerIndexSignal.map { return ($0 as Int == RegistrationIndex.ConfirmVC.toInt()) }
+        
+        RAC(confirmButton, "hidden") <~ indexIsConfirmSignal.notEach()
+        RAC(flowView, "highlightedIndex") <~ registerIndexSignal
 
-        let details = self.fulfilmentNav().bidDetails
+        RAC(registrationNetworkModel, "createNewUser") <~ RACObserve(self, "createNewUser")
+        RAC(registrationNetworkModel, "details") <~ RACObserve(self, "details")
+        
+        details = self.fulfilmentNav().bidDetails
         flowView.details = details
         bidDetailsPreviewView.bidDetails = details
         
@@ -44,10 +59,33 @@ class RegisterViewController: UIViewController {
 
         if let subscribableVC = controller as? RegistrationSubController {
             subscribableVC.finishedSignal.subscribeCompleted({ [weak self] () -> Void in
-                self?.flowView.update()
                 self?.goToNextVC()
-                return
+                self?.flowView.update()
             })
+        }
+    }
+
+    @IBAction func confirmTapped(sender: AnyObject) {
+        confirmButton.enabled = false
+        
+        registrationNetworkModel.completedSignal.subscribeNext({ [weak self] (_) -> Void in
+
+            self?.moveToNextViewController()
+            return
+            
+        }, error: { [weak self] (error) -> Void in
+            println("Error with registrationNetworkModel create or update")
+        })
+        
+        registrationNetworkModel.start()
+    }
+    
+    func moveToNextViewController() {
+        if createNewUser {
+            self.fulfilmentNav().parentViewController?.dismissViewControllerAnimated(true, completion: nil)
+            
+        } else {
+            self.performSegue(.ConfirmRegistrationandBid)
         }
     }
 }