+++ Kiosk/Storyboards/StoryboardIdentifiers.swift
@@ -3,15 +3,18 @@
 public enum SegueIdentifier : String {
     case ArtsyUserHasNotRegisteredCard = "Artsy User Has Not Registered Card"
     case ConfirmBid = "Confirm Bid"
+    case ConfirmRegistrationandBid = "Confirm Registration and Bid"
     case ConfirmyourBidArtsyLogin = "Confirm your Bid - Artsy Login"
     case ConfirmyourBidBidderFound = "Confirm your Bid - Bidder Found"
     case ConfirmyourBidBidderNotFound = "Confirm your Bid - Bidder Not Found"
     case EmailLoginConfirmedHighestBidder = "Email Login Confirmed Highest Bidder"
+    case LoadAdminWebViewController = "Load Admin Web View Controller"
     case LoadWebView = "Load Web View"
     case LoggedinWithNewUser = "Logged in With New User "
     case PINConfirmed = "PIN Confirmed"
     case RegisterCreditCardToPostal = "RegisterCreditCardToPostal"
     case RegisterPostalToConfirm = "RegisterPostalToConfirm"
+    case ShowAdminOptions = "Show Admin Options"
     case SubmittedanEmailforUserDetails = "Submitted an Email for User Details"
 }